[pages:Home] Move `app/Home` directory to `pages/` directory
[tsconfig.json] Add path: `pages/`
[webpack-config:paths] Add path: `pages`
[webpack-config:base] Add alias for `pages`
[jest.config.js] Add module mapping: `pages`, add `pages/` dir to config
for `App Tests` project
[.eslintrc.js] Add alias mapping for `pages`
