+++ app/Routes/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Routes as RouterRoutes } from 'react-router-dom';
 
-import Home from 'app/Home';
+import Home from 'pages/Home';
 
 export default function Routes() {
   return (
