+++ tsconfig.json
@@ -13,7 +13,8 @@
     "allowSyntheticDefaultImports": true,
     "forceConsistentCasingInFileNames": true,
     "paths": {
-      "app/*": ["./app/*"]
+      "app/*": ["./app/*"],
+      "pages/*": ["./pages/*"]
     }
   }
 }
